Bug Fix - Circuitscape ascii input cell size
An error in the way ArcGIS rounds cell size values when converting
raster to ascii was causing rasters with cell sizes that include more
than 12 decimal places to produce ascii files with slightly different
cell sizes. This created problems with Circuitscape.

This fix rounds the arc cell size environment variable to a maximum of
12 decimal places before ascii conversion is attempted.